chore: improve RadioButtonGroup exports
